requirements.txt
streamlit
phidata
duckduckgo-search
groq
openai
exa_py
